Split controllers
